+++ resources-tests/laravelInsight/annotation/ColumnWithoutAnnotationInspection.npe.php
@@ -1,0 +1,11 @@
+<?php
+
+use Illuminate\Database\Eloquent\Model;
+
+class BaseModel extends Model {
+    protected $timestamps = true;
+    protected $primaryKey = 'id';
+}
+
+class ShouldBeRemovedProgramatically extends BaseModel {
+}